Create unified (uses fallback) resolver, renamed unified->dynamic
